work in progress features

+++ src/main/java/com/github/ringlocker/substancecraft/client/shader/ShaderEffectTicker.java
@@ -62,12 +62,12 @@ public class ShaderEffectTicker {
         builder.putFloat(1.0F + ((PlayerEffectState.strength(SubstanceCraftEffects.COLOR_ENHANCEMENT) + 1) * 0.2F));
         builder.putFloat(23F - (PlayerEffectState.strength(SubstanceCraftEffects.COLOR_RESOLUTION) + 1));
         builder.putFloat((float) getTime());
-        builder.putFloat(0.05F * (PlayerEffectState.strength(SubstanceCraftEffects.DYNAMIC_COLOR) + 1));
-        builder.putFloat(1.0F + (0.2F * (float) (PlayerEffectState.strength(SubstanceCraftEffects.MOSAIC) + 1)));
-        builder.putFloat(0.33F * (float) (PlayerEffectState.strength(SubstanceCraftEffects.SURFACE_WARP) + 1));
-        builder.putFloat(PlayerEffectState.strength(SubstanceCraftEffects.DOUBLE_VISION) + 1);
-        builder.putFloat(0.005f * (1 + Math.max(8, PlayerEffectState.strength(SubstanceCraftEffects.DOUBLE_VISION))));
-        builder.putFloat(1.1f + (Math.max(8, PlayerEffectState.strength(SubstanceCraftEffects.DOUBLE_VISION))) / 6.66f);
+        builder.putFloat(0.05F * (PlayerEffectState.strength(SubstanceCraftEffects.DYNAMIC_COLOR) + 1.0F));
+        builder.putFloat(1.0F + (0.2F * (PlayerEffectState.strength(SubstanceCraftEffects.MOSAIC) + 1.0F)));
+        builder.putFloat(0.33F * (PlayerEffectState.strength(SubstanceCraftEffects.SURFACE_WARP) + 1.0F));
+        builder.putFloat(PlayerEffectState.strength(SubstanceCraftEffects.DOUBLE_VISION) + 1.0F);
+        builder.putFloat(0.005f * (1.0F + Math.max(8.0F, PlayerEffectState.strength(SubstanceCraftEffects.DOUBLE_VISION))));
+        builder.putFloat(1.1f + (Math.max(8.0F, PlayerEffectState.strength(SubstanceCraftEffects.DOUBLE_VISION))) / 6.66f);
     }
 
     private static int getTime() {
